requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
